requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.8.0
django-filter==21.1
python-decouple==3.4
psycopg2-binary==2.9.1
Faker==8.12.1
coverage==5.5
pytest==6.2.5
pytest-django==4.4.0